Реализована фильтрация

+++ designsite/designpro/views.py
@@ -4,7 +4,7 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate
 from django.contrib.auth.forms import AuthenticationForm
 from django.shortcuts import render, redirect
-from .forms import RegistrationForm, DesignRequestForm
+from .forms import RegistrationForm, DesignRequestForm, DesignRequestFilterForm
 from django.contrib.auth.models import User
 from django.contrib.auth import login
 from .models import CustomUser
@@ -35,7 +35,6 @@ def registration(request):
             user.first_name = full_name
             user.save()
 
-            # Создаем объект CustomUser и связываем его с пользователем
             custom_user, created = CustomUser.objects.get_or_create(user=user, defaults={'full_name': full_name})
 
             login(request, user)
@@ -81,12 +80,21 @@ def create_design_request(request):
 
     return render(request, 'autorized/create_request.html', {'form': form})
 
-def my_design_requests(request):
 
-    my_requests = DesignRequest.objects.filter(user=request.user.customuser)
+def my_design_requests(request):
+    custom_user = CustomUser.objects.get(user=request.user)
+    my_requests = DesignRequest.objects.filter(user=custom_user)
+
+    filter_form = DesignRequestFilterForm(request.GET)
+
+    if filter_form.is_valid():
+        status = filter_form.cleaned_data['status']
+        if status:
+            my_requests = my_requests.filter(status=status)
 
     context = {
         'my_requests': my_requests,
+        'filter_form': filter_form,
     }
 
     return render(request, 'autorized/my_design_requests.html', context)
@@ -101,23 +109,20 @@ def request_detail(request, request_id):
     return render(request, 'autorized/request_detail.html', context)
 
 @login_required
-def delete_design_request(request, request_id):
-    # Получаем заявку
-    design_request = get_object_or_404(DesignRequest, pk=request_id)
+def delete_request(request, pk):
+    design_request = get_object_or_404(DesignRequest, pk=pk)
 
-    # Проверяем, что текущий пользователь - автор заявки
     if design_request.user.user != request.user:
         raise Http404("Вы не можете удалить эту заявку.")
 
-    # Удаляем заявку
+    if design_request.status in ["Принято в работу", "Выполнено"]:
+        raise Http404("Вы не можете удалить заявку со статусом \"Принято в работу\" или \"Выполнено\".")
+
     design_request.delete()
 
     return redirect('my_design_requests')
 
-def delete_request(request, pk):
-    design_request = DesignRequest.objects.get(pk=pk)
-    design_request.delete()
-    return redirect('my_design_requests')
+
